Model Cleanup, UI Stuff
Shifting focus to how I want to build the front-end (using fake data) and
working out the data-models. Then I'll go back to the backend to work on
the caching and indexing.

Created a new controller for serving the main "index" (read landing) page
for the applicaiton. Also modified the existing main and index views to not
be the play default pages. Added in twitter bootstrap and jquer from the
CDN as well.