Prototype: Project Service + Basic Controller
Just getting the feel of what it is like to build with services and repositories.
Made a basic controller for projects that reads and creates. There is a bit of work
to do with the data model regarding what fields are required and which ones we can
live without (like projects not having any version upon creation).

Also, definitely need to be using DI. I ended up initializing it all in the controllers
companion object to avoid the Play framework creating a new isntance of the controller
(and thus my memory stores) on each request. I can live with the controller being
instantiated each time, but we need to use DI to make the repo / service a singleton.