DI Support (Scaldi) - Basic Setup
Nothing fancy at the moment, but at least I got Scaldi working with my application.
I did bump up to Play 2.4.0 which require some changes around Global. Need to do a
bit of additional work to get rid of global now that it's all deprecated and what
not.

Fort this I created a couple of repo modules and registered them via the application.conf.
I also edited the existing api.ProjectController to use DI instead of the static
definition withing the singleton. Now that this is working, future work can just _use_
it for free (or mostly free).